+++ R/smooth_data.R
@@ -23,11 +23,7 @@
 #' Replaces data with a rolling median value. The `k` parameter is used to control the number of points
 #' to include in the median calculation, should be odd but will convert to an odd number with a warning.
 #'
-#' - **savgol**: Based on the `savgol` function from the `pracma` package.
-#' Applies a Savitzky–Golay filter which fits a fourth order polynomial to a sliding range of data.
-#' The `k` parameter is used controls the number of points in the window.
-#'
-#' - **savgol**: Based on the `dlmSmooth` function from the `dlm` package.
+#' - **kalman**: Based on the `dlmSmooth` function from the `dlm` package.
 #' Applies a Kalman filter which process model to the data.
 #' The `k` parameter is used controls the amount of smoothing.
 #'
@@ -37,24 +33,25 @@
 #' ggplot2::geom_line(color="black", na.rm=TRUE) +
 #' ggplot2::geom_line(data=smoothed, color="darkred", na.rm=TRUE)
 apply_smoothing <- function(data, y_var, method, index=NULL, k=7){
-  stopifnot(method %in% c("rollmean", "rollmedian", "savgol", "kalman"))
+  stopifnot(method %in% c("rollmean", "rollmedian", "kalman"))
 
   #get indices if not specified
   if(is.null(index)){index <- data$Index}
 
+  #if number not odd and should be make odd
+  if(k %% 2 == 0){k <- k + 1}
+  if(method == "savgol" & k <3){k <- 3} #k can't be 1 and must be odd integer
+
   #if using rolling methods that have NA's at the start, grab a little extra
   if(method %in% c("rollmean", "rollmedian")){
     newstart <- ifelse(min(index)-k < 1, 1, min(index)-k)
-    newend <- ifelse(max(index)+k > length(index), max(index), max(index)+k)
+    newend <- ifelse(max(index)+(k*2) > length(index), max(index), max(index)+(k*2))
     smooth_idx <- seq(from=newstart, to =newend, by=1)
   }else{
     smooth_idx <- index
   }
 
-  #if number not odd and should be make odd
-  if(k %% 2 == 0){k <- k + 1}
-
-  #pull out just data to smooth with a little extra for starting and ending to prevent unneeded NA's
+    #pull out just data to smooth with a little extra for starting and ending to prevent unneeded NA's
     #get wider index if a subset of data, if not outside the bounds
     #index should directly map to rows, but in case....
     subdat <- data %>% filter(.data$Index %in% smooth_idx)
@@ -62,16 +59,22 @@ apply_smoothing <- function(data, y_var, method, index=NULL, k=7){
 
   #apply smoothing method
   if(method == "rollmean"){
-    filled <- rollmean(vals, na.pad=TRUE, k=k)
+    filled <- rollapply(vals, width = k, FUN = mean, na.rm = TRUE, fill = NA)
   }
 
   if(method == "rollmedian"){
-    filled <- rollmedian(vals, na.pad=TRUE, k=k)
+    filled <- rollapply(vals, width = k, FUN = median, na.rm = TRUE, fill = NA)
+
   }
 
-  if(method == "savgol"){
-    filled <- savgol(vals, k, forder = 4, dorder = 0)
-  }
+  # if(method == "savgol"){
+  #   #fill NA's so function will work
+  #   vals <- zoo::na.approx(vals, na.rm = FALSE)
+  #   vals <- zoo::na.locf(vals, na.rm = FALSE)        # forward fill
+  #   vals <- zoo::na.locf(vals, fromLast = TRUE)      # backward fill
+  #
+  #   filled <- savgol(vals, k, forder = 4, dorder = 0)
+  # }
 
   if(method == "kalman"){
     filled <- dlmSmooth(vals, dlmModPoly(1, dV = 20, dW = 1/as.numeric(k)))$s[-1]